refactor: 💡 update directory structures
