+++ src/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import EntryRoute from './presentation/routes';
+import EntryRoute from './pages';
 import GlobalStyles from './styles/GlobalStyles';
 import * as serviceWorker from './serviceWorker';
